refactor: concrete effect
